requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-environ
django-filter
django-cors-headers
psycopg2-binary
drf-spectacular
Pillow

celery
redis

stripe
paystackapi
mpesa-api  

twilio
